feat: (GDPR compliance) provide informed consent for loading embedded youtube videos that place identifying tracking data in the users browser
we use youtube-nocookie.com to load yt videos. this is misleading as
this stills saves identifying data in browser local storage (hence "no
cookie"), which is still against the privacy regulations of the EU and
UK. we now display an informational message over the img placeholder of
the iframe (before any iframe is loaded) informing the user that
interaction via clicking to play the video will result in Google placing
these disguised cookies.